refactor(tooling): split root task config
