+++ tools/moon/scripts/postprocess_types.mbtx
@@ -1,4 +1,18 @@
 ///|
+/// Post-processes TypeScript declarations generated from Vize's JSON Schema.
+///
+/// The schema stays the source of truth for the public config shape, but a few
+/// TypeScript-facing details cannot be expressed precisely by the current
+/// schema-to-types pipeline:
+///
+/// - `RegExp` inputs are runtime values, not JSON Schema values.
+/// - Public lint preset and rule metadata should be exposed as narrow string
+///   unions instead of being inferred from generic schema fragments.
+/// - `$schema` helps editor JSON files, but it is not useful on the TypeScript
+///   configuration object that users import from the package.
+///
+/// Keeping this as a small MoonBit script lets generation stay inside the
+/// repository toolchain while avoiding another TypeScript-only maintenance path.
 import {
   "moonbitlang/core/env",
   "moonbitlang/x@0.4.41/fs" @xfs,
@@ -7,17 +21,25 @@ import {
 }
 
 ///|
+/// Joins two path fragments using MoonBit's path implementation instead of
+/// manual string concatenation.
 fn join_path(base : String, child : String) -> String {
   let path : @path.Path = base
   path.join(child).to_string()
 }
 
 ///|
+/// Checks path existence through a tiny wrapper so repository discovery reads
+/// like a sequence of domain checks rather than low-level filesystem calls.
 fn safe_path_exists(path : String) -> Bool {
   @xfs.path_exists(path)
 }
 
 ///|
+/// Reads a UTF-8 file and collapses IO failures into `None`.
+///
+/// The CLI entry point owns user-facing diagnostics, so helper functions keep
+/// the control flow compact and leave error wording in one place.
 fn read_file_to_string(path : String) -> String? {
   try {
     Some(@xfs.read_file_to_string(path))
@@ -27,6 +49,11 @@ fn read_file_to_string(path : String) -> String? {
 }
 
 ///|
+/// Writes a complete file and reports success as a boolean.
+///
+/// The generated declaration is rewritten atomically from the caller's point of
+/// view: either the replacement content was accepted by the filesystem helper,
+/// or the script exits with a clear failure message.
 fn write_string_to_file(path : String, content : String) -> Bool {
   try {
     @xfs.write_string_to_file(path, content)
@@ -37,6 +64,11 @@ fn write_string_to_file(path : String, content : String) -> Bool {
 }
 
 ///|
+/// Walks upward from `start` until it finds the Vize repository root.
+///
+/// Two markers are checked to avoid accidentally treating a nested fixture or
+/// package directory as the root. Tests can bypass this search by setting
+/// `VIZE_REPO_ROOT` to an isolated fixture workspace.
 fn find_repo_root(start : String) -> String? {
   let current : @path.Path = start
   let current_path = current.resolve().to_string()
@@ -53,11 +85,19 @@ fn find_repo_root(start : String) -> String? {
 }
 
 ///|
+/// Returns the generated TypeScript declaration path relative to the repository
+/// root.
 fn generated_types_path(root : String) -> String {
   join_path(root, "npm/vize/src/types/generated.ts")
 }
 
 ///|
+/// Applies Vize-specific declaration patches that are intentionally kept out of
+/// the generic JSON Schema generation step.
+///
+/// The transform is idempotent. That property matters because release and local
+/// generation tasks may be rerun after a partial failure; repeated execution
+/// should converge on the same file rather than stacking duplicate aliases.
 fn postprocess_generated_types(content : String) -> String {
   let content = content.replace_all(old="MuseaA11YConfig", new="MuseaA11yConfig")
   let content = content.replace_all(
@@ -92,6 +132,12 @@ fn postprocess_generated_types(content : String) -> String {
 }
 
 ///|
+/// Main command implementation for generated config type post-processing.
+///
+/// The function resolves the repository root, reads the generated declaration,
+/// applies the deterministic declaration patches, and writes the result back to
+/// disk. Returning an exit code keeps `main` trivial and makes tests exercise the
+/// same control flow as the CLI invocation.
 fn run_app() -> Int {
   let repo_root = match @sys.get_env_var("VIZE_REPO_ROOT") {
     Some(value) => {
@@ -125,6 +171,7 @@ fn run_app() -> Int {
 }
 
 ///|
+/// Script entry point.
 fn main {
   @sys.exit(run_app())
 }
